re-enable auto-fetch
I was recording https://reallifemag.com with browsertrix-crawler and
noticed that the images failed to render on replay. This was because the
images that browsertrix-crawler's browser fetched during archiving
weren't the same responsive images that my browser requested during
replay.

I noticed that the pywb configuration that browsertrix-crawler
is using has `enable-auto-fetch` turned off due to being commented out. Was
this intentional? Should we re-enable to improve the fidelity of
archives created with browsertrix-crawler and cloud?