Update README.md, intro to regular expressions and fix for confusing line.
Better intro to regular expressions, clarification of confusing and self-contradicting line "the scope include regexes will crawl all page URLs that match example.com/(crawl-this|crawl-that), but skip URLs that end with 'skip-me'" since "skip" means it won't be excluded but it can also be interpreted as it will be excluded.

If this is sent as a pull request, I didn't mean to, I only wanted to save changes on my fork. Consider rejecting, there is more work to be done.