Disable disk utilization check by default (#850)
Related to https://github.com/webrecorder/browsertrix-crawler/issues/848

Several users have had issues with disk utilization checks, including
the values reported by `df` inside the crawler container having
unexpected results for mounted volumes. The commonly recommended
solution to this is to use `docker system ps`, but that is of course not
available within the Docker container itself.

This PR changes disk utilization checks to be an opt-in feature by
setting the default value to `0` (disabled).